Support forked tests using new Anvil process
